+++ game/world.go
@@ -69,12 +69,13 @@ func (w *World) IsBlocked(x, y int) bool {
 
 // MoveCost returns the movement cost to enter the tile at (x, y).
 // Derived from MoveCooldownFor so pathfinding cost stays in sync with movement speed.
-func (w *World) MoveCost(x, y int) int {
+// Always >= 1.0 so the A* Manhattan heuristic in geom.FindPath remains admissible.
+func (w *World) MoveCost(x, y int) float64 {
 	t := w.TileAt(x, y)
 	if t == nil {
 		return 1
 	}
-	return int(MoveCooldownFor(t) / defaultMoveCooldown)
+	return float64(MoveCooldownFor(t)) / float64(defaultMoveCooldown)
 }
 
 // noGrowRadius is the Euclidean radius around the spawn point and any structure
